Fix: [OSX] support custom url handler

+++ src/nw_content.cc
@@ -940,7 +940,7 @@ bool ApplicationShouldHandleReopenHook(bool hasVisibleWindows) {
 
 void OSXOpenURLsHook(const std::vector<GURL>& startup_urls) {
   scoped_ptr<base::ListValue> arguments(new base::ListValue());
-  for (int i = 0; i < startup_urls.size(); i++)
+  for (size_t i = 0; i < startup_urls.size(); i++)
     arguments->AppendString(startup_urls[i].spec());
   SendEventToApp("nw.App.onOpen", arguments.Pass());
 }
